requirements.txt
Django>=4.2,<5.0
python-decouple>=3.8
dj-database-url>=2.1.0
python-dotenv>=1.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
psycopg2-binary>=2.9.9
django-filter>=23.5
# Exportación de reportes
openpyxl>=3.1.2
reportlab>=4.0.7
# Integración climática
requests>=2.31.0
